WIP: THIS BREAKS MORE TESTS
I'm out of time, it's quarter past nine and I've a child still up.

This more or less makes things work from the perspective of just trying
everything out in the browser. It breaks all the tests (largely because
they depend on Steps::PARAMS, which has changed to accommodate more than
one researcher)